Add possibility to set command parameter as String.
This fix establishes the desired behavior for the "command" parameter as
described in the icinga2 docs and in the icinga2 module parameter
description:

"The command. This can either be an array of individual command arguments.
Alternatively a string can be specified in which case the shell interpreter
(usually /bin/sh) takes care of parsing the command."

Update the rspec tests.