Temporarily add back $zone and $endpoint.
Re-add deprecated $zone and $endpoint parameters to allow for a transition period, especially for users using virtual or exported resources.
In fact, with exported resources in storeconfigs that use the old parameters, all puppet runs will fail when upgrading the module to a version
that only has the new parameters. Having a transition period will allow puppet to run and export new resources to storeconfig that use the new parameter names.