Fix dependency issues with apt repository
This is needed as there is currently no dependency between the
icinga2 packages and Class['apt::update']. This can lead to problems
where the package is installed before apt knowns of the repository.

It also removes the anchor usage of the icinga2::repo class and moves
the dependency management to icinga::repo itself. This is needed as
other modules might also add apt repositories and require
Class['apt::update']. When using the anchor pattern this breaks as
it introduces a dependency cycle.